put viewModel in separate file
